csi: Publish Filesystem volumes using bind mounts
Kubelet recursively changes gids on volumes when the Pod SecurityContext
fsGroup setting is specified. This doesn't work as expected on KubeSAN's
symlinks to the mounted staging directory. Kubelet changes the gid on
the symlink itself and does not recurse into the Filesystem volume's
contents.

Use a bind mount when publishing Filesystem volumes so that Kubelet can
change gids correctly when fsGroup is specified.

+++ internal/csi/node/publish.go
@@ -7,6 +7,8 @@ import (
 	"errors"
 	"os"
 
+	"k8s.io/mount-utils"
+
 	"github.com/container-storage-interface/spec/lib/go/csi"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -33,17 +35,32 @@ func (s *NodeServer) NodePublishVolume(ctx context.Context, req *csi.NodePublish
 		return nil, err
 	}
 
-	// create symlink to device for Kubernetes (replace if there is a file, empty dir, or symlink in its place;
-	// Kubernetes places an empty directory at the path where block volumes should be staged/published)
+	if req.VolumeCapability.GetMount() != nil {
+		// create bind mount for Filesystem volumes
 
-	err := os.Remove(req.TargetPath)
-	if err != nil && !errors.Is(err, os.ErrNotExist) {
-		return nil, err
-	}
+		err := os.MkdirAll(req.TargetPath, 0750)
+		if err != nil {
+			return nil, err
+		}
 
-	err = os.Symlink(req.StagingTargetPath, req.TargetPath)
-	if err != nil {
-		return nil, err
+		err = s.mounter.Mount(req.StagingTargetPath, req.TargetPath, "none", []string{"bind"})
+		if err != nil {
+			_ = os.Remove(req.TargetPath)
+			return nil, err
+		}
+	} else {
+		// create symlink to device for Kubernetes (replace if there is a file, empty dir, or symlink in its place;
+		// Kubernetes places an empty directory at the path where block volumes should be staged/published)
+
+		err := os.Remove(req.TargetPath)
+		if err != nil && !errors.Is(err, os.ErrNotExist) {
+			return nil, err
+		}
+
+		err = os.Symlink(req.StagingTargetPath, req.TargetPath)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	// success
@@ -64,11 +81,21 @@ func (s *NodeServer) NodeUnpublishVolume(ctx context.Context, req *csi.NodeUnpub
 		return nil, status.Errorf(codes.InvalidArgument, "must specify target path")
 	}
 
-	// remove symlink created by NodePublishVolume
+	isMountPoint, err := s.mounter.IsMountPoint(req.TargetPath)
+	if err == nil && isMountPoint {
+		// remove bind mount created by NodePublishVolume for a Filesystem volume
 
-	err := os.Remove(req.TargetPath)
-	if err != nil && !errors.Is(err, os.ErrNotExist) {
-		return nil, err
+		err := mount.CleanupMountPoint(req.TargetPath, s.mounter, true)
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		// remove symlink created by NodePublishVolume for a Block volume
+
+		err := os.Remove(req.TargetPath)
+		if err != nil && !errors.Is(err, os.ErrNotExist) {
+			return nil, err
+		}
 	}
 
 	// success
